requirements.txt
six
pyjwt
requests
urllib3
certifi
